911961# Refactor Visual Recognition connectors in one class to avoid code duplication
- Create VisualRecognitionService class
- Move implementation from CreateClassifier, ClassifyImage and DetectFaces to VisualRecognitionService
